# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CONEXT API",
        "version": "0.3"
    },
    "paths": {
...
        "api/admin/business-types/{id}": {
            "delete": {
                "tags": [
                    "Business"
                ],
                "summary": "Delete business-types",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\BusinessTypeController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The business-types that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Business type not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/categories/{id}": {
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete category",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\CategoryController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The category that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/contact/{id}": {
            "delete": {
                "tags": [
                    "Contact"
                ],
                "summary": "Delete contact",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\ContactController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The contact that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Contact not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/events/{id}": {
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Delete events",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\EventController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The event that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/expertises/{id}": {
            "delete": {
                "tags": [
                    "Expertises"
                ],
                "summary": "Delete expertises",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\ExpertiseController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The expertises that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Expertise not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/explore/{id}": {
            "delete": {
                "tags": [
                    "Explore"
                ],
                "summary": "Delete explore",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\ExploreController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The explore that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Explore not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/hot-needs/{id}": {
            "delete": {
                "tags": [
                    "Hot-needs"
                ],
                "summary": "Delete hot-needs",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\HotNeedController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The hot-needs that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "HotNeed not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/local-conexts/{id}": {
            "delete": {
                "tags": [
                    "Local-conexts"
                ],
                "summary": "Delete local-conexts",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\LocalConextController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The local-conexts that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Local conext not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/users/{id}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete users",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\UserController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The users that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Users not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "api/admin/vouchers/{id}": {
            "delete": {
                "tags": [
                    "Vouchers"
                ],
                "summary": "Delete vouchers",
                "description": "This can only be done by the logged in user.",
                "operationId": "App\\Http\\Controllers\\VoucherController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The vouchers that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Voucher not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}